backend/requirements.txt
fastapi==0.95.0
langchain==0.1.5
pinecone_client==2.2.2
pydantic==1.10.12
python-dotenv==1.0.0
uvicorn==0.21.1
python-multipart==0.0.6
openai==0.27.8
pypdf==3.12.0
tiktoken==0.4.0
psycopg2-binary~=2.9.3
faiss-cpu==1.7.4
sentence-transformers
pdfminer
pydub==0.25.1
retry==0.9.2
httpx==0.25.2
databases==0.8.0
asyncpg==0.29.0

